fix: delete dynamic lang function cause translation change from cookies
