--- wealth-sync/src/app/[locale]/layout.tsx
@@ -1,54 +1,0 @@
-import { CookieConsentDialog } from "@/components/CookiesConsent/CookiesConsent";
-import { Navbar } from "@/components/Navbar/Navbar";
-import { Providers } from "@/components/Providers";
-import { PlatformConnectionProvider } from "@/lib/contexts/PlatformConnectionContext";
-import "@/styles/globals.css";
-import { NextIntlClientProvider } from "next-intl";
-import { type Metadata } from "next";
-import { Geist } from "next/font/google";
-import { getMessages } from "next-intl/server";
-import { hasLocale } from "next-intl";
-import { notFound } from "next/navigation";
-import { routing } from "../../i18n/routing";
-
-export const metadata: Metadata = {
-  title: "Wealth Sync",
-  description: "Wealth Calculator App that helps you manage your finances",
-  icons: [{ rel: "icon", url: "/dollar-app-logo.jpeg" }],
-};
-
-const geist = Geist({
-  subsets: ["latin"],
-  variable: "--font-geist-sans",
-});
-
-export default async function RootLayout({
-  children,
-  params,
-}: Readonly<{
-  children: React.ReactNode;
-  params: Promise<{ locale: string }>;
-}>) {
-  const messages = await getMessages();
-  const { locale } = await params;
-
-  if (!hasLocale(routing.locales, locale)) {
-    notFound();
-  }
-
-  return (
-    <html lang="en" className={`${geist.variable}`} suppressHydrationWarning>
-      <body>
-        <Providers>
-          <NextIntlClientProvider messages={messages}>
-            <Navbar />
-            <PlatformConnectionProvider>
-              {children}
-              <CookieConsentDialog />
-            </PlatformConnectionProvider>
-          </NextIntlClientProvider>
-        </Providers>
-      </body>
-    </html>
-  );
-}
