+++ wealth-sync/src/app/auth/layout.tsx
