+++ wealth-sync/src/app/auth/register/page.tsx
