+++ wealth-sync/src/app/coming-soon/layout.tsx
