+++ wealth-sync/src/app/dashboard/layout.tsx
