+++ wealth-sync/src/app/dashboard/page.tsx
