+++ wealth-sync/src/app/layout.tsx
@@ -21,7 +21,11 @@ const geist = Geist({
 
 export default async function RootLayout({
   children,
-}: Readonly<{ children: React.ReactNode }>) {
+  params,
+}: Readonly<{
+  children: React.ReactNode;
+  params: Promise<{ locale: string }>;
+}>) {
   const messages = await getMessages();
 
   return (
