+++ wealth-sync/src/app/page.tsx
