+++ wealth-sync/src/app/pricing/page.tsx
