+++ wealth-sync/src/app/privacy/page.tsx
