+++ wealth-sync/src/app/profile/page.tsx
