+++ wealth-sync/src/app/terms/page.tsx
